Create an outline for database with change notifiers
I think this idea of a "model" that automatically updates the views is
flawed.  Essentially, that is my React mindset and that doesn't work
because I need to be able to keep state.

I can imagine that this is possible, but that would require tracking the
widgets in the model, which doesn't appear to be what these models are
for.