Create outline for database the third time
Each of the "entities" has a list of versions where the last version is
the most recent one.  We can call 'update' on this entity to append a
new version and to emit the 'onChanged' signal.

The database keeps a list of all the entities, currently, we manually
join the tables.  That could be done more efficiently.