Work: Start implementing the 'BillItemDelegate' the intended way
You are not supposed to put proper widgets there for some reason.  It
seems to be possible though workarounds with 'setIndexWidget', however,
that does not seem to be intended.

This feels odd, since something very similar is possible with
'createEditor'.  Reconciliation is also not an issue because of the
stuff we store in the model indices.

Maybe it is difficult to render widgets in other locations constantly?
But that is essentially what layouts are all about...

I don't get it.