Work: Make saving operations failable and async
Somehow, this isn't working, but I don't really understand why.

When I launch the application, we successfully download the database but
then it doesn't return 'QDialog::DialogCode::Accepted'.  Thus we crash
when we assert that value.

My theory is, that I am doing some stuff on another thread that needs to
happen on the UI thread.