Never start application if database not loaded
If the program isn't able to load the database and the user closes the
retry dialog, we simply quit the application.

I tried doing this before but this didn't work, because
'QApplication::quit()' doesn't put an event on the queue if the event
loop hasn't been started.  That seems very unexpected and should be
fixed by them.